requirements.txt
python-telegram-bot==20.6
python-dateutil
holidays
pandas
rapidfuzz